+++ public/app/plugins/datasource/cloudwatch/partials/query.parameter.html
@@ -66,6 +66,7 @@
 				<li>{{namespace}}</li>
 				<li>{{region}}</li>
 				<li>{{period}}</li>
+				<li>{{label}}</li>
 				<li>{{YOUR_DIMENSION_NAME}}</li>
 			</ul>
 		</info-popover>
